+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Release Notes
 
+## Unreleased
+- Improved multi-site support when elements are not enabled for all sites. ([#50](https://github.com/craftcms/gatsby-source-craft/issues/50))
+
 ## 2.0.3 - 2021-08-19
 - Updated gatsby-graphql-source-toolkit to 2.x.